refactored unit tests into a BaseTest class and each Test class inherited for each CQL measure
