+++ src/test/resources/bmi001/bundles/expected.json
@@ -280,7 +280,7 @@
         {
           "resource": {
             "resourceType": "Observation",
-            "id": "bmi",
+            "id": "bmi-1",
             "meta": {
               "profile": [
                 "http://hl7.org/fhir/StructureDefinition/vitalsigns"
@@ -613,7 +613,7 @@
         {
           "resource": {
             "resourceType": "Observation",
-            "id": "bmi",
+            "id": "bmi-2",
             "meta": {
               "profile": [
                 "http://hl7.org/fhir/StructureDefinition/vitalsigns"
